% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{This contains the regression coefficents of linreg function.}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the regression coefficents of linreg function.
}
