% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Generic method for getting the fitted values}
\usage{
pred(p, ...)

\method{pred}{linreg}(p, ...)
}
\arguments{
\item{p}{An object}

\item{...}{Further arguments passed to or from other methods}

\item{object}{An object of linreg class}
}
\value{
This prints out predicted or fitted values for the object passed.
}
\description{
Generic method for getting the fitted values

This contains the fitted values of linreg function.
}
