% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{This contains the summary method for linreg function.}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the summary method for linreg function.
}
